Fixed problem with the temp set panel being displayed when in Fan or Off mode when the page is initially loaded/refreshed.
Removed temp formatting from the logic and moved it to the display (index.html)

Changed time/date refresh from making a call to the server to get the date/time to using the browser date/time.

Added the last known temp next to sensor name in the selector panel

Added detection for when the relay controller has failed by logging the when the controller was last successfully read.
Added a failed controller panel to the display status bar when the controller has failed.

Changed the WiFi status check interval to every 10 seconds from 5 seconds

Added notes to the setup instructions to install the mysql connector library and to set the max_connections for mySQL